+++ lib/src/transport.dart
@@ -1,29 +1,41 @@
 import 'dart:async';
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:meta/meta.dart';
 import 'package:protobuf/protobuf.dart';
 
 import 'codec.dart';
+import 'error.dart' as centrifuge;
 import 'proto/client.pb.dart' hide Error;
 
 typedef Transport TransportBuilder({
   @required String url,
-  @required Map<String, dynamic> headers,
+  @required TransportConfig config,
 });
 
 typedef Future<WebSocket> WebSocketBuilder();
 
+class TransportConfig {
+  TransportConfig(
+      {this.pingInterval = const Duration(seconds: 25),
+      this.headers = const <String, dynamic>{}});
+
+  final Duration pingInterval;
+  final Map<String, dynamic> headers;
+}
+
 Transport protobufTransportBuilder(
-    {@required String url, @required Map<String, dynamic> headers}) {
+    {@required String url, @required TransportConfig config}) {
   final replyDecoder = ProtobufReplyDecoder();
   final commandEncoder = ProtobufCommandEncoder();
 
   final transport = Transport(
     () => WebSocket.connect(
-          url,
-          headers: headers,
-        ),
+      url,
+      headers: config.headers,
+    ),
+    config,
     commandEncoder,
     replyDecoder,
   );
@@ -31,21 +43,34 @@ Transport protobufTransportBuilder(
   return transport;
 }
 
-class Transport {
-  Transport(this._socketBuilder, this._commandEncoder, this._replyDecoder);
+abstract class GeneratedMessageSender {
+  Future<Rep>
+      sendMessage<Req extends GeneratedMessage, Rep extends GeneratedMessage>(
+          Req request, Rep result);
+}
+
+class Transport implements GeneratedMessageSender {
+  Transport(this._socketBuilder, this._config, this._commandEncoder,
+      this._replyDecoder);
 
   final WebSocketBuilder _socketBuilder;
   WebSocket _socket;
   final CommandEncoder _commandEncoder;
   final ReplyDecoder _replyDecoder;
+  final TransportConfig _config;
 
-  Future open(void onPush(Push push), {Function onError, void onDone()}) async {
+  Future open(void onPush(Push push),
+      {Function onError,
+      void onDone(String reason, bool shouldReconnect)}) async {
     _socket = await _socketBuilder();
+    if (_config.pingInterval != Duration.zero) {
+      _socket.pingInterval = _config.pingInterval;
+    }
 
     _socket.listen(
       _onData(onPush),
       onError: onError,
-      onDone: onDone,
+      onDone: _onDone(onDone),
     );
   }
 
@@ -53,10 +78,11 @@ class Transport {
 
   final _completers = <int, Completer<GeneratedMessage>>{};
 
-  Future<Rep> send<Req extends GeneratedMessage, Rep extends GeneratedMessage>(
-      Req request, Rep result) async {
+  @override
+  Future<Rep>
+      sendMessage<Req extends GeneratedMessage, Rep extends GeneratedMessage>(
+          Req request, Rep result) async {
     final command = _createCommand(request);
-
     final reply = await _sendCommand(command);
 
     final filledResult = _processResult(result, reply);
@@ -78,6 +104,11 @@ class Transport {
     _completers[command.id] = completer;
 
     final data = _commandEncoder.convert(command);
+
+    if (_socket == null) {
+      throw centrifuge.ClientDisconnectedError;
+    }
+
     _socket.add(data);
 
     return completer.future;
@@ -85,7 +116,7 @@ class Transport {
 
   T _processResult<T extends GeneratedMessage>(T result, Reply reply) {
     if (reply.hasError()) {
-      throw reply.error;
+      throw centrifuge.Error.custom(reply.error.code, reply.error.message);
     }
     result.mergeFromBuffer(reply.result);
     return result;
@@ -103,9 +134,26 @@ class Transport {
         return MethodType.SUBSCRIBE;
       case HistoryRequest:
         return MethodType.HISTORY;
+      case RPCRequest:
+        return MethodType.RPC;
       default:
         throw ArgumentError('unknown request type');
     }
+  }
+
+  Function _onDone(void Function(String, bool) onDone) {
+    return () {
+      String reason;
+      bool reconnect = true;
+      if (_socket.closeReason != null) {
+        try {
+          final Map<String, dynamic> info = jsonDecode(_socket.closeReason);
+          reason = info['reason'];
+          reconnect = info['reconnect'] ?? true;
+        } catch (_) {}
+      }
+      onDone(reason, reconnect);
+    };
   }
 
   Function _onData(void onPush(Push push)) {