+++ CHANGELOG.md
@@ -1,3 +1,28 @@
+## [0.5.1]
+* Skip handling for events from server-side subscriptions resulted in null pointer dereference 
+
+## [0.5.0]
+* Update protobuf dependency to `^1.0.1`, thanks [@Holofox](https://github.com/Holofox)
+
+## [0.4.1]
+* Implement `removeSubscription` method, thanks [@tiamo](https://github.com/tiamo)
+
+## [0.4.0]
+* **Breaking changes** `connected` property removed
+
+## [0.3.0]
+* **Breaking changes** Changed API to avoid returning futures where they do not make sense – in `connect` and `subscribe` methods.
+* Client now uses Websocket Ping/Pong frames to find broken connection
+* Support for private channel subscription using `onPrivateSub` configuration callback function
+* Fix several null pointer dereferences in edge cases, throw `ClientDisconnectedError` when there is an attempt to send protocol request over non-connected client.
+* Fix examples where subscribe could happen before stream handlers set
+
+## [0.2.0]
+* **Breaking changes** Replaced subscribe method with getSubscription in Client. 
+* Added reconnection.
+* Added history method to Subscription. Thanks [vanelizarov](https://github.com/vanelizarov).
+* Added rpc method.
+
 ## [0.1.0]
 * Added authorization with JWT
 * Added subscription to private channel with JWT